refactor(step3): separate racing car game class and application
